+++ erbildaphne.comMvcWebApp/Controllers/GNewsController.cs
@@ -8,6 +8,7 @@ using Newtonsoft.Json;
 using System.Net.Http.Headers;
 using System.Net.Http.Json;
 using System.Reflection;
+using System.Security.Claims;
 using System.Text;
 using static System.Net.WebRequestMethods;
 
@@ -49,40 +50,70 @@ namespace erbildaphne.comMvcWebApp.Controllers
         public async Task<IActionResult> List()
         {
             var token = HttpContext.Session.GetString("JWTToken");
+
+            if (token == null)
+            {
+                return RedirectToAction("Login", "Account");
+            }
+
+
             var http = _httpClientFactory.CreateClient("Client");
             http.DefaultRequestHeaders.Authorization = new AuthenticationHeaderValue(JwtBearerDefaults.AuthenticationScheme, token);
+            var roleClaims = User.Claims.Where(c => c.Type == ClaimTypes.Role && c.Value == "Editor");
 
-            var result = await http.GetAsync("gNews");
-            if (result.StatusCode == System.Net.HttpStatusCode.OK)
+            if (roleClaims != null)
             {
-                var jsonData = await result.Content.ReadAsStringAsync();
-                var data = JsonConvert.DeserializeObject<List<GNewsViewModel>>(jsonData);
-                //Json Serialize işlemleri için NewtonSoft paketini yüklüyoruz.
-                return View(data);
+                var result = await http.GetAsync("gNews");
+                if (result.StatusCode == System.Net.HttpStatusCode.OK)
+                {
+                    var jsonData = await result.Content.ReadAsStringAsync();
+                    var data = JsonConvert.DeserializeObject<List<GNewsViewModel>>(jsonData);
+                    //Json Serialize işlemleri için NewtonSoft paketini yüklüyoruz.
+                    return View(data);
+                }
+                else
+                {
+                    // API'den başarısız bir yanıt geldiğinde hata sayfasını göster
+                    return View("Error");
+                }
             }
-            else
-            {
-                // API'den başarısız bir yanıt geldiğinde hata sayfasını göster
-                return View("Error");
-            }
+
+
+            return RedirectToAction("Login", "Account");
+
+
         }
 
 
 
         [HttpGet]
-        //[Authorize(Roles = "admin")]
         public async Task<IActionResult> Create()
         {
             var token = HttpContext.Session.GetString("JWTToken");
+
+            if (token == null)
+            {
+                return RedirectToAction("Login", "Account");
+            }
+
+
             var http = _httpClientFactory.CreateClient("Client");
             http.DefaultRequestHeaders.Authorization = new AuthenticationHeaderValue(JwtBearerDefaults.AuthenticationScheme, token);
+            var roleClaims = User.Claims.Where(c => c.Type == ClaimTypes.Role && c.Value == "Editor");
 
-            return View(new GNewsViewModel()); // Burada GNewsViewModel nesnesi geçiriliyor.
+            if (roleClaims != null)
+            {
+                return View(new GNewsViewModel());
+            }
+
+
+            return RedirectToAction("Login", "Account");
+
+            // Burada GNewsViewModel nesnesi geçiriliyor.
         }
 
 
         [HttpPost]
-        //[Authorize(Roles = "admin")]
         public async Task<IActionResult> Create(GNewsViewModel model, IFormFile image)
         {
 
@@ -120,26 +151,42 @@ namespace erbildaphne.comMvcWebApp.Controllers
             try
             {
                 var token = HttpContext.Session.GetString("JWTToken");
+
+                if (token == null)
+                {
+                    return RedirectToAction("Login", "Account");
+                }
+
+
                 var http = _httpClientFactory.CreateClient("Client");
                 http.DefaultRequestHeaders.Authorization = new AuthenticationHeaderValue(JwtBearerDefaults.AuthenticationScheme, token);
+                var roleClaims = User.Claims.Where(c => c.Type == ClaimTypes.Role && c.Value == "Editor");
 
-                var jsonContent = JsonConvert.SerializeObject(model);
-
-                var content = new StringContent(jsonContent, Encoding.UTF8, "application/json");
-
-                var result = await http.PostAsync("gNews", content);
-                var error = result.Content.ReadAsStringAsync();               
-
-                if (result.IsSuccessStatusCode)
+                if (roleClaims != null)
                 {
-                    // Başarılı POST işleminden sonra yapılacak işlemler
-                    return RedirectToAction("List");
+                    var jsonContent = JsonConvert.SerializeObject(model);
+
+                    var content = new StringContent(jsonContent, Encoding.UTF8, "application/json");
+
+                    var result = await http.PostAsync("gNews", content);
+                    var error = result.Content.ReadAsStringAsync();
+
+                    if (result.IsSuccessStatusCode)
+                    {
+                        // Başarılı POST işleminden sonra yapılacak işlemler
+                        return RedirectToAction("List");
+                    }
+                    else
+                    {
+                        // API'den başarısız bir yanıt geldiğinde hata sayfasını göster
+                        return RedirectToAction("List");
+                    }
                 }
-                else
-                {
-                    // API'den başarısız bir yanıt geldiğinde hata sayfasını göster
-                    return RedirectToAction("List");
-                }
+
+
+                return RedirectToAction("Login", "Account");
+
+
 
 
 
@@ -151,11 +198,6 @@ namespace erbildaphne.comMvcWebApp.Controllers
             }
         }
 
-        
-
-        
-       
-        
 
 
 
@@ -181,72 +223,104 @@ namespace erbildaphne.comMvcWebApp.Controllers
         }
 
         [HttpGet]
-        //[Authorize(Roles = "admin")]
+
         public async Task<IActionResult> Publish(int id)
         {
             var token = HttpContext.Session.GetString("JWTToken");
-            var http = _httpClientFactory.CreateClient("Client");
-            http.DefaultRequestHeaders.Authorization = new AuthenticationHeaderValue(JwtBearerDefaults.AuthenticationScheme, token);
 
-            var gNewsResult = await http.GetAsync("gNews/" + id);
-
-            if (gNewsResult.StatusCode == System.Net.HttpStatusCode.OK)
+            if (token == null)
             {
-                var gNewsJsonData = await gNewsResult.Content.ReadAsStringAsync();
-                var gNewsData = JsonConvert.DeserializeObject<GNewsViewModel>(gNewsJsonData);
-
-                // Yayın durumunu değiştir
-                gNewsData.IsPublished = !gNewsData.IsPublished;
-
-                // Güncellenmiş veriyi JSON'a dönüştür
-                var jsonContent = JsonConvert.SerializeObject(gNewsData);
-                var content = new StringContent(jsonContent, Encoding.UTF8, "application/json");
-
-                // API'ye güncellenmiş veriyi gönder
-                var updateResult = await http.PutAsync("gNews/" + id.ToString(), content);
-                var error = updateResult.Content.ReadAsStringAsync();
-
-                return RedirectToAction("List");
-
+                return RedirectToAction("Login", "Account");
             }
 
-            // Hata durumunda
-            return View("Error");
+
+            var http = _httpClientFactory.CreateClient("Client");
+            http.DefaultRequestHeaders.Authorization = new AuthenticationHeaderValue(JwtBearerDefaults.AuthenticationScheme, token);
+            var roleClaims = User.Claims.Where(c => c.Type == ClaimTypes.Role && c.Value == "Editor");
+
+            if (roleClaims != null)
+            {
+                var gNewsResult = await http.GetAsync("gNews/" + id);
+
+                if (gNewsResult.StatusCode == System.Net.HttpStatusCode.OK)
+                {
+                    var gNewsJsonData = await gNewsResult.Content.ReadAsStringAsync();
+                    var gNewsData = JsonConvert.DeserializeObject<GNewsViewModel>(gNewsJsonData);
+
+                    // Yayın durumunu değiştir
+                    gNewsData.IsPublished = !gNewsData.IsPublished;
+
+                    // Güncellenmiş veriyi JSON'a dönüştür
+                    var jsonContent = JsonConvert.SerializeObject(gNewsData);
+                    var content = new StringContent(jsonContent, Encoding.UTF8, "application/json");
+
+                    // API'ye güncellenmiş veriyi gönder
+                    var updateResult = await http.PutAsync("gNews/" + id.ToString(), content);
+                    var error = updateResult.Content.ReadAsStringAsync();
+
+                    return RedirectToAction("List");
+
+                }
+
+                // Hata durumunda
+                return View("Error");
+            }
+
+
+            return RedirectToAction("Login", "Account");
+
+
         }
 
         [HttpGet]
-        //[Authorize(Roles = "admin")]
+
         public async Task<IActionResult> Edit(int id)
         {
             var token = HttpContext.Session.GetString("JWTToken");
+
+            if (token == null)
+            {
+                return RedirectToAction("Login", "Account");
+            }
+
+
             var http = _httpClientFactory.CreateClient("Client");
             http.DefaultRequestHeaders.Authorization = new AuthenticationHeaderValue(JwtBearerDefaults.AuthenticationScheme, token);
+            var roleClaims = User.Claims.Where(c => c.Type == ClaimTypes.Role && c.Value == "Editor");
 
-            var gNewsResult = await http.GetAsync("gNews/" + id);
-
-
-            if (gNewsResult.StatusCode == System.Net.HttpStatusCode.OK)
+            if (roleClaims != null)
             {
-                var gNewsJsonData = await gNewsResult.Content.ReadAsStringAsync();
-                var gNewsData = JsonConvert.DeserializeObject<GNewsViewModel>(gNewsJsonData);
+                var gNewsResult = await http.GetAsync("gNews/" + id);
 
-                return View(gNewsData);
+
+                if (gNewsResult.StatusCode == System.Net.HttpStatusCode.OK)
+                {
+                    var gNewsJsonData = await gNewsResult.Content.ReadAsStringAsync();
+                    var gNewsData = JsonConvert.DeserializeObject<GNewsViewModel>(gNewsJsonData);
+
+                    return View(gNewsData);
+                }
+                else
+                {
+                    return View("Error");
+                }
             }
-            else
-            {
-                return View("Error");
-            }
+
+
+            return RedirectToAction("Login", "Account");
+
+
         }
-        
+
 
         [HttpPost]
-        //[Authorize(Roles = "admin")]
-        public async Task<IActionResult> Edit(int id, GNewsViewModel model,IFormFile? image)
-        {            
-             
+
+        public async Task<IActionResult> Edit(int id, GNewsViewModel model, IFormFile? image)
+        {
+
             if (id == model.Id)
             {
-                if(image != null)
+                if (image != null)
                 {
                     var uploadPath = Path.Combine(Directory.GetCurrentDirectory(), "wwwroot\\uploads\\gnews");
 
@@ -269,26 +343,42 @@ namespace erbildaphne.comMvcWebApp.Controllers
 
 
                 var token = HttpContext.Session.GetString("JWTToken");
+
+                if (token == null)
+                {
+                    return RedirectToAction("Login", "Account");
+                }
+
+
                 var http = _httpClientFactory.CreateClient("Client");
                 http.DefaultRequestHeaders.Authorization = new AuthenticationHeaderValue(JwtBearerDefaults.AuthenticationScheme, token);
+                var roleClaims = User.Claims.Where(c => c.Type == ClaimTypes.Role && c.Value == "Editor");
 
-                var gNewsResult = await http.GetAsync("gNews/" + id);
-                var existing = await gNewsResult.Content.ReadAsStringAsync();
-                var existingData = JsonConvert.DeserializeObject<GNewsViewModel>(existing);
-                model.TotalSource = model.LeaningLeft + model.LeaningRight + model.LeaningCenter;
-                existingData.Title = model.Title;
-                existingData.Description = model.Description;
-                existingData.Content = model.Content;
-                existingData.LeaningLeft = model.LeaningLeft;
-                existingData.LeaningRight = model.LeaningRight;
-                existingData.LeaningCenter = model.LeaningCenter;
-                existingData.TotalSource= model.TotalSource;
+                if (roleClaims != null)
+                {
+                    var gNewsResult = await http.GetAsync("gNews/" + id);
+                    var existing = await gNewsResult.Content.ReadAsStringAsync();
+                    var existingData = JsonConvert.DeserializeObject<GNewsViewModel>(existing);
+                    model.TotalSource = model.LeaningLeft + model.LeaningRight + model.LeaningCenter;
+                    existingData.Title = model.Title;
+                    existingData.Description = model.Description;
+                    existingData.Content = model.Content;
+                    existingData.LeaningLeft = model.LeaningLeft;
+                    existingData.LeaningRight = model.LeaningRight;
+                    existingData.LeaningCenter = model.LeaningCenter;
+                    existingData.TotalSource = model.TotalSource;
 
-                var jsonContent = JsonConvert.SerializeObject(existingData);
-                var content = new StringContent(jsonContent, Encoding.UTF8, "application/json");
-                var result = await http.PutAsync("gNews/" + id.ToString(), content);
-                var error = result.Content.ReadAsStringAsync();
-                return RedirectToAction("List");
+                    var jsonContent = JsonConvert.SerializeObject(existingData);
+                    var content = new StringContent(jsonContent, Encoding.UTF8, "application/json");
+                    var result = await http.PutAsync("gNews/" + id.ToString(), content);
+                    var error = result.Content.ReadAsStringAsync();
+                    return RedirectToAction("List");
+                }
+
+
+                return RedirectToAction("Login", "Account");
+
+
 
 
 
@@ -298,56 +388,86 @@ namespace erbildaphne.comMvcWebApp.Controllers
         }
 
         [HttpGet]
-        //[Authorize(Roles = "admin")]
         public async Task<IActionResult> Delete(int id)
         {
             var token = HttpContext.Session.GetString("JWTToken");
+
+            if (token == null)
+            {
+                return RedirectToAction("Login", "Account");
+            }
+
+
             var http = _httpClientFactory.CreateClient("Client");
             http.DefaultRequestHeaders.Authorization = new AuthenticationHeaderValue(JwtBearerDefaults.AuthenticationScheme, token);
+            var roleClaims = User.Claims.Where(c => c.Type == ClaimTypes.Role && c.Value == "Editor");
 
-            var gNewsResult = await http.GetAsync("gNews/" + id);
-            var existing = await gNewsResult.Content.ReadAsStringAsync();
-            var existingData = JsonConvert.DeserializeObject<GNewsViewModel>(existing);
-            if (existingData.IsDeleted)
+            if (roleClaims != null)
             {
-                var result = await http.DeleteAsync("gNews/" + id.ToString());
-                var error = result.Content.ReadAsStringAsync();
-                return RedirectToAction("List");
-            }
-            else
-            {
+                var gNewsResult = await http.GetAsync("gNews/" + id);
+                var existing = await gNewsResult.Content.ReadAsStringAsync();
+                var existingData = JsonConvert.DeserializeObject<GNewsViewModel>(existing);
+                if (existingData.IsDeleted)
+                {
+                    var result = await http.DeleteAsync("gNews/" + id.ToString());
+                    var error = result.Content.ReadAsStringAsync();
+                    return RedirectToAction("List");
+                }
+                else
+                {
 
-                existingData.IsDeleted = true;
-                var jsonContent = JsonConvert.SerializeObject(existingData);
-                var content = new StringContent(jsonContent, Encoding.UTF8, "application/json");
-                var result2 = await http.PutAsync("gNews/" + id.ToString(), content);
-                var error = result2.Content.ReadAsStringAsync();
-                return RedirectToAction("List");
+                    existingData.IsDeleted = true;
+                    var jsonContent = JsonConvert.SerializeObject(existingData);
+                    var content = new StringContent(jsonContent, Encoding.UTF8, "application/json");
+                    var result2 = await http.PutAsync("gNews/" + id.ToString(), content);
+                    var error = result2.Content.ReadAsStringAsync();
+                    return RedirectToAction("List");
+                }
             }
+
+
+            return RedirectToAction("Login", "Account");
+
 
         }
-        
+
         [HttpGet]
-        //[Authorize(Roles = "admin")]
+
         public async Task<IActionResult> RemoveDelete(int id)
         {
             var token = HttpContext.Session.GetString("JWTToken");
+
+            if (token == null)
+            {
+                return RedirectToAction("Login", "Account");
+            }
+
+
             var http = _httpClientFactory.CreateClient("Client");
             http.DefaultRequestHeaders.Authorization = new AuthenticationHeaderValue(JwtBearerDefaults.AuthenticationScheme, token);
+            var roleClaims = User.Claims.Where(c => c.Type == ClaimTypes.Role && c.Value == "Editor");
 
-            var gNewsResult = await http.GetAsync("gNews/" + id);
-            var existing = await gNewsResult.Content.ReadAsStringAsync();
-            var existingData = JsonConvert.DeserializeObject<GNewsViewModel>(existing);
-            if (existingData.IsDeleted)
+            if (roleClaims != null)
             {
-                existingData.IsDeleted = false;
-                var jsonContent = JsonConvert.SerializeObject(existingData);
-                var content = new StringContent(jsonContent, Encoding.UTF8, "application/json");
-                var result2 = await http.PutAsync("gNews/" + id.ToString(), content);
-                var error = result2.Content.ReadAsStringAsync();
+                var gNewsResult = await http.GetAsync("gNews/" + id);
+                var existing = await gNewsResult.Content.ReadAsStringAsync();
+                var existingData = JsonConvert.DeserializeObject<GNewsViewModel>(existing);
+                if (existingData.IsDeleted)
+                {
+                    existingData.IsDeleted = false;
+                    var jsonContent = JsonConvert.SerializeObject(existingData);
+                    var content = new StringContent(jsonContent, Encoding.UTF8, "application/json");
+                    var result2 = await http.PutAsync("gNews/" + id.ToString(), content);
+                    var error = result2.Content.ReadAsStringAsync();
+                    return RedirectToAction("List");
+                }
                 return RedirectToAction("List");
             }
-            return RedirectToAction("List");
+
+
+            return RedirectToAction("Login", "Account");
+
+
         }
